feat: adapt code organization for alpaca calls
