+++ libs/coin-modules/coin-stellar/package.json
@@ -51,6 +51,7 @@
     "@faker-js/faker": "^8.4.1",
     "@types/invariant": "^2.2.2",
     "@types/jest": "^29.5.10",
+    "@types/node": "^16.11.7",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.1"
   },
