+++ libs/coin-modules/coin-stellar/src/bridge/broadcast.ts
@@ -1,7 +1,7 @@
 import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
 import type { AccountBridge, Operation, SignedOperation } from "@ledgerhq/types-live";
-import { broadcastTransaction as apiBroadcast } from "./network";
-import { Transaction } from "./types";
+import { broadcastTransaction as apiBroadcast } from "../network";
+import { Transaction } from "../types";
 
 /**
  * Broadcast a signed transaction
