+++ libs/coin-modules/coin-stellar/src/bridge/deviceTransactionConfig.ts
@@ -1,6 +1,6 @@
 import type { AccountLike, Account } from "@ledgerhq/types-live";
 import type { CommonDeviceTransactionField as DeviceTransactionField } from "@ledgerhq/coin-framework/transaction/common";
-import type { Transaction, TransactionStatus } from "./types";
+import type { Transaction, TransactionStatus } from "../types";
 
 export type ExtraDeviceTransactionField =
   | {
