+++ libs/coin-modules/coin-stellar/src/bridge/tokens.ts
@@ -5,7 +5,7 @@ import type { SyncConfig, TokenAccount } from "@ledgerhq/types-live";
 import { parseCurrencyUnit } from "@ledgerhq/coin-framework/currencies/parseCurrencyUnit";
 import { encodeOperationId } from "@ledgerhq/coin-framework/operation";
 import { findTokenById, listTokensForCryptoCurrency } from "@ledgerhq/cryptoassets";
-import type { BalanceAsset, StellarOperation } from "./types";
+import type { BalanceAsset, StellarOperation } from "../types";
 
 export const getAssetIdFromTokenId = (tokenId: string): string => tokenId.split("/")[2];
