+++ libs/coin-modules/coin-stellar/src/config.ts
@@ -1,18 +1,9 @@
-import { CurrencyConfig, CoinConfig } from "@ledgerhq/coin-framework/config";
-import { MissingCoinConfig } from "@ledgerhq/coin-framework/errors";
+import buildCoinConfig, { type CurrencyConfig } from "@ledgerhq/coin-framework/config";
 
-export type StellarCoinConfig = CurrencyConfig;
+export type StellarConfig = Record<string, never>;
 
-let coinConfig: CoinConfig<StellarCoinConfig> | undefined;
+export type StellarCoinConfig = CurrencyConfig & StellarConfig;
 
-export function setCoinConfig(config: CoinConfig<StellarCoinConfig>): void {
-  coinConfig = config;
-}
+const coinConfig = buildCoinConfig<StellarCoinConfig>();
 
-export function getCoinConfig(): StellarCoinConfig {
-  if (!coinConfig) {
-    throw new MissingCoinConfig();
-  }
-
-  return coinConfig();
-}
+export default coinConfig;
