+++ libs/coin-modules/coin-stellar/src/index.ts
@@ -1,0 +1,1 @@
+export { createBridges } from "./bridge";
