+++ libs/coin-modules/coin-stellar/src/signer/getAddress.ts
@@ -2,9 +2,9 @@ import { GetAddressFn } from "@ledgerhq/coin-framework/bridge/getAddressWrapper"
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import { GetAddressOptions } from "@ledgerhq/coin-framework/derivation";
 import { StrKey } from "@stellar/stellar-sdk";
-import { StellarSigner } from "./types/signer";
+import { StellarSigner } from "../types/signer";
 
-function resolver(signerContext: SignerContext<StellarSigner>): GetAddressFn {
+function getAddress(signerContext: SignerContext<StellarSigner>): GetAddressFn {
   return async (deviceId: string, { path, verify }: GetAddressOptions) => {
     const rawPublicKey = await signerContext(deviceId, async signer => {
       const { rawPublicKey } = await signer.getPublicKey(path, verify);
@@ -21,4 +21,4 @@ function resolver(signerContext: SignerContext<StellarSigner>): GetAddressFn {
   };
 }
 
-export default resolver;
+export default getAddress;
