+++ libs/coin-modules/coin-stellar/src/test/cli.ts
@@ -1,8 +1,8 @@
 import invariant from "invariant";
 import type { AccountLike, Account, AccountLikeArray } from "@ledgerhq/types-live";
 import { getAccountCurrency } from "@ledgerhq/coin-framework/account/helpers";
-import type { Transaction } from "./types";
-import { getAssetIdFromTokenId } from "./tokens";
+import type { Transaction } from "../types";
+import { getAssetIdFromTokenId } from "../bridge/tokens";
 
 const options = [
   {
