+++ libs/coin-modules/coin-stellar/src/types/errors.ts
